pspmix: Fix signed unsigned comparison

+++ plugins/pspmix/pspmixserver.c
@@ -1825,7 +1825,7 @@ bool pspmix_server_init(char *nspace, pmix_rank_t rank, const char *clusterid,
 #if PMIX_VERSION_MAJOR >= 4
     if (mset(PSPMIX_LOG_INFOARR)) {
 	mlog("%s: PMIx_server_init info:\n", __func__);
-	for (int j = 0; j < cbdata.ninfo; j++) {
+	for (size_t j = 0; j < cbdata.ninfo; j++) {
 	    char * istr = PMIx_Info_string(&cbdata.info[j]);
 	    mlog("%s\n", istr);
 	    free(istr);
